+++ src/collective/gridlets/__init__.py
@@ -1,7 +1,3 @@
 from zope.i18nmessageid import MessageFactory
 
 _ = MessageFactory("collective.gridlets")
-
-
-def initialize(context):
-    """Initializer called when used as a Zope 2 product."""