+++ src/collective/gridlets/behavior.py
@@ -1,23 +1,22 @@
-from rwproperty import getproperty, setproperty
-from zope.component import adapts
-from zope.interface import implements, alsoProvides
+from zope.interface import alsoProvides
 from zope import schema
 from z3c.form.interfaces import IEditForm
 
-from plone.directives import form
-from plone.dexterity.interfaces import IDexterityContent
+from plone.supermodel import model
+from plone.autoform import directives as form
+from plone.autoform.interfaces import IFormFieldProvider
 
 from collective.gridlets import _
-from collective.gridlets.widget import GridsterField
+from collective.gridlets.widget import GridsterFieldWidget
 
 
-class IGridlets(form.Schema):
+class IGridlets(model.Schema):
     """Add gridlets serialization field to content
     """
 
     form.omitted('gridlets')
     form.no_omit(IEditForm, 'gridlets')
-    form.widget(gridlets=GridsterField)
+    form.widget(gridlets=GridsterFieldWidget)
     gridlets = schema.TextLine(
         title=_(u"Gridlets"),
         description=_(u"gridlets_description"),
@@ -25,19 +24,4 @@ class IGridlets(form.Schema):
         required=False,
     )
 
-alsoProvides(IGridlets, form.IFormFieldProvider)
-
-
-class Gridlets(object):
-    implements(IGridlets)
-    adapts(IDexterityContent)
-
-    def __init__(self, context):
-        self.context = context
-
-    @getproperty
-    def gridlets(self):
-        return self.context.gridlets
-    @setproperty
-    def gridlets(self, value):
-        self.context.gridlets = value
+alsoProvides(IGridlets, IFormFieldProvider)